chore(macro): snapshot testing
