+++ packages/babel-plugin-lingui-macro/test/fixtures/jsx-keep-forced-newlines.expected.js
@@ -1,2 +1,2 @@
 import { Trans as _Trans } from "@lingui/react"
-;<_Trans id={"<stripped>"} message={"Keep multiple\nforced\nnewlines!"} />
+;<_Trans id={"9xE5pD"} message={"Keep multiple\nforced\nnewlines!"} />
