+++ packages/babel-plugin-lingui-macro/test/index.ts
@@ -1,108 +1,9 @@
-import fs from "fs"
 import path from "path"
-import {
-  PluginObj,
-  transformFileSync,
-  TransformOptions,
-  transformSync,
-} from "@babel/core"
-import prettier from "prettier"
-// use package path instead relative because we want
-// to test it in from /dist folder in integration tests
-import linguiMacroPlugin, {
-  LinguiPluginOpts,
-} from "@lingui/babel-plugin-lingui-macro"
-import {
-  JSXAttribute,
-  jsxExpressionContainer,
-  JSXIdentifier,
-  stringLiteral,
-} from "@babel/types"
-import { NodePath } from "@babel/traverse"
-
-export type TestCase = {
-  name?: string
-  input?: string
-  expected?: string
-  filename?: string
-  production?: boolean
-  useTypescriptPreset?: boolean
-  macroOpts?: LinguiPluginOpts
-  /** Remove hash id from snapshot for more stable testing */
-  stripId?: boolean
-  only?: boolean
-  skip?: boolean
-}
-
-const testCases: Record<string, TestCase[]> = {
-  "js-t": require("./js-t").default,
-  "js-plural": require("./js-plural").default,
-  "js-select": require("./js-select").default,
-  "js-selectOrdinal": require("./js-selectOrdinal").default,
-  "jsx-trans": require("./jsx-trans").default,
-  "jsx-select": require("./jsx-select").default,
-  "jsx-plural": require("./jsx-plural").default,
-  "jsx-selectOrdinal": require("./jsx-selectOrdinal").default,
-  "js-defineMessage": require("./js-defineMessage").default,
-  "js-useLingui": require("./js-useLingui").default,
-}
-
-function stripIdPlugin(): PluginObj {
-  return {
-    visitor: {
-      JSXOpeningElement: (path) => {
-        const idAttr = path
-          .get("attributes")
-          .find(
-            (attr) =>
-              attr.isJSXAttribute() &&
-              (attr.node.name as JSXIdentifier).name === "id"
-          ) as NodePath<JSXAttribute>
-
-        if (idAttr) {
-          idAttr
-            .get("value")
-            .replaceWith(jsxExpressionContainer(stringLiteral("<stripped>")))
-        }
-      },
-    },
-  }
-}
+import { transformSync } from "@babel/core"
+import { getDefaultBabelOptions } from "./macroTester"
 
 describe("macro", function () {
   process.env.LINGUI_CONFIG = path.join(__dirname, "lingui.config.js")
-
-  const getDefaultBabelOptions = (
-    transformType: "plugin" | "macro" = "plugin",
-    macroOpts: LinguiPluginOpts = {},
-    isTs: boolean = false,
-    stripId = false
-  ): TransformOptions => {
-    return {
-      filename: "<filename>" + (isTs ? ".tsx" : "jsx"),
-      configFile: false,
-      babelrc: false,
-      presets: [],
-      plugins: [
-        "@babel/plugin-syntax-jsx",
-        transformType === "plugin"
-          ? [linguiMacroPlugin, macroOpts]
-          : [
-              "macros",
-              {
-                lingui: macroOpts,
-                // macro plugin uses package `resolve` to find a path of macro file
-                // this will not follow jest pathMapping and will resolve path from ./build
-                // instead of ./src which makes testing & developing hard.
-                // here we override resolve and provide correct path for testing
-                resolvePath: (source: string) => require.resolve(source),
-              },
-            ],
-
-        ...(stripId ? [stripIdPlugin] : []),
-      ],
-    }
-  }
 
   const transformTypes = ["plugin", "macro"] as const
 
@@ -131,97 +32,6 @@ describe("macro", function () {
         throw e
       }
     }
-
-  Object.keys(testCases).forEach((suiteName) => {
-    describe(`${suiteName}`, () => {
-      const cases = testCases[suiteName]
-
-      const clean = (value: string) =>
-        prettier.format(value, { parser: "babel-ts" }).replace(/\n+/, "\n")
-
-      cases.forEach(
-        (
-          {
-            name,
-            input,
-            expected,
-            filename,
-            production,
-            useTypescriptPreset,
-            only,
-            skip,
-            macroOpts,
-            stripId,
-          },
-          index
-        ) => {
-          let group = describe
-          if (only) group = describe.only
-          if (skip) group = describe.skip
-          group(name != null ? name : `${suiteName} #${index + 1}`, () => {
-            transformTypes.forEach((transformType) => {
-              it(transformType, () => {
-                const babelOptions = getDefaultBabelOptions(
-                  transformType,
-                  macroOpts,
-                  useTypescriptPreset,
-                  stripId
-                )
-                expect(input || filename).toBeDefined()
-
-                const originalEnv = process.env.NODE_ENV
-
-                if (production) {
-                  process.env.NODE_ENV = "production"
-                }
-
-                if (useTypescriptPreset) {
-                  babelOptions.presets.push("@babel/preset-typescript")
-                }
-
-                try {
-                  if (filename) {
-                    const inputPath = path.relative(
-                      process.cwd(),
-                      path.join(__dirname, "fixtures", filename)
-                    )
-                    const expectedPath = inputPath.replace(
-                      /\.js$/,
-                      ".expected.js"
-                    )
-                    const expected = fs
-                      .readFileSync(expectedPath, "utf8")
-                      .replace(/\r/g, "")
-                      .trim()
-
-                    const _babelOptions = {
-                      ...babelOptions,
-                      cwd: path.dirname(inputPath),
-                    }
-
-                    const actual = transformFileSync(inputPath, _babelOptions)
-                      .code.replace(/\r/g, "")
-                      .trim()
-                    expect(clean(actual)).toEqual(clean(expected))
-                  } else {
-                    const actual = transformSync(
-                      input,
-                      babelOptions
-                    ).code.trim()
-
-                    expect(clean(actual)).toEqual(clean(expected))
-                  }
-                } finally {
-                  process.env.LINGUI_CONFIG = ""
-                  process.env.NODE_ENV = originalEnv
-                }
-              })
-            })
-          })
-        }
-      )
-    })
-  })
 
   it("Should throw error if used without babel-macro-plugin", async () => {
     await expect(async () => {
