+++ packages/babel-plugin-lingui-macro/test/js-selectOrdinal.test.ts
@@ -1,0 +1,16 @@
+import { macroTester } from "./macroTester"
+
+macroTester({
+  cases: [
+    {
+      code: `
+        import { t, selectOrdinal } from '@lingui/core/macro'
+        t\`This is my \${selectOrdinal(count, {
+          one: "#st",
+          "two": \`#nd\`,
+          other: ("#rd")
+        })} cat\`
+      `,
+    },
+  ],
+})
